Rename analyzer module to ascot; refresh wasm for eager class parsing
Renames server/src/analyzer -> server/src/ascot (plus the wasm artifact,
build script, and every import) now that the module is unambiguously
ascot-specific. Also refreshes the bundled wasm to pick up ascot's
eager-on-open class parsing, which keeps queryCls/queryMem from paying
for a whole-workspace parse burst on whichever request asks first.
