+++ server/src/providers/inlayHint.ts
@@ -1,11 +1,11 @@
 import { InlayHint, InlayHintParams } from "vscode-languageserver/node";
-import * as analyzer from "../analyzer";
+import * as ascot from "../ascot";
 
 /**
  * Handler function for the `textDocument/inlayHint` request.
  */
 export async function onInlayHint(params: InlayHintParams): Promise<InlayHint[]> {
-	const hints = await analyzer.inlayHint(params.textDocument.uri, params.range);
+	const hints = await ascot.inlayHint(params.textDocument.uri, params.range);
 
 	return hints.map((hint) => ({
 		position: hint.position,
