+++ server/src/providers/signatureHelp.ts
@@ -24,7 +24,7 @@ import {
 } from "../utils/functions";
 import { ServerSpec, SignatureHelpDocCache, SignatureHelpMacroContext } from "../utils/types";
 import { documents } from "../utils/variables";
-import { getAnalyzedClassMember } from "../analyzer";
+import { getAnalyzedClassMember } from "../ascot";
 import * as ld from "../utils/languageDefinitions";
 import { prettifyNormalArg } from "./hover";
 import { URI } from "vscode-uri";
