+++ server/src/utils/functions.ts
@@ -32,7 +32,7 @@ import systemVariables from "../documentation/systemVariables.json";
 
 // Initialize turndown and tune it for Documatic HTML
 import { default as TurndownService } from "turndown";
-import { ClassInfo, getAnalyzedClasses } from "../analyzer";
+import { ClassInfo, getAnalyzedClasses } from "../ascot";
 const turndown = new TurndownService({
 	codeBlockStyle: "fenced",
 	blankReplacement: (content, node: HTMLElement) => (node.nodeName == "SPAN" ? node.outerHTML : ""),
